FIX linter errors
